refactor(state): split usePatient into session and editor hooks
