+++ .gitignore
@@ -60,7 +60,3 @@ __pycache__/
 *.pyc
 package-lock.old
 package-lock.old
-QWEN.md
-
-# Local notes
-/docs/backlog.md
